finalizado

+++ NossoQueijo.Repositorio/RepositoriosEF/UsuarioRepositorio.cs
@@ -21,15 +21,33 @@ namespace NossoQueijo.Repositorio.RepositoriosEF
         {
             return _contexto.Usuarios
                 .Include(x => x.TipoUsuario)
-                .Include(x => x.Endereco)
-                .ThenInclude(y => y.Cidade)
-                .ThenInclude(y => y.Estado)
+                .Include(x => x.Enderecos)
+                .ThenInclude(x => x.Cidade)
+                .ThenInclude(x => x.Estado)
+                .Include(x => x.FichasProducao)
                 .ToList();
+        }
+
+        public void AdicionarPersonalizado(Usuario usuario)
+        {
+            usuario.TipoUsuario = _contexto.TiposUsuario
+                .First(x => x.idTipoUsuario == usuario.TipoUsuario.idTipoUsuario);
+            _contexto.Usuarios.Add(usuario);
+            _contexto.SaveChanges();
+        }
+
+        public void AtualizarPersonalizado(Usuario usuario)
+        {
+            usuario.TipoUsuario = _contexto.TiposUsuario
+                .First(x => x.idTipoUsuario == usuario.TipoUsuario.idTipoUsuario);
+            _contexto.Usuarios.Update(usuario);
+            _contexto.SaveChanges();
         }
 
         public Usuario BuscarPorId(int id)
         {
             return _contexto.Usuarios
+                .Include(x => x.TipoUsuario)
                 .First(x => x.idUsuario == id);
         }
 
@@ -37,9 +55,9 @@ namespace NossoQueijo.Repositorio.RepositoriosEF
         {
             return _contexto.Usuarios
                 .Include(x => x.TipoUsuario)
-                .Include(x => x.Endereco)
-                .ThenInclude(y => y.Cidade)
-                .ThenInclude(y => y.Estado)
+                .Include(x => x.Enderecos)
+                .ThenInclude(x => x.Cidade)
+                .ThenInclude(x => x.Estado)
                 .Where(x => x.TipoUsuario.idTipoUsuario == idTipoUsuario)
                 .ToList();
         }
@@ -50,30 +68,47 @@ namespace NossoQueijo.Repositorio.RepositoriosEF
                 .First(x => (x.Email == email) && (x.Senha == senha));
         }
 
-        public bool RemoverPorIdTipoUsuario(int idTipoUsuario)
+        public void RemoverPorIdTipoUsuario(int idTipoUsuario)
         {
             IEnumerable<Usuario> usuariosRemover = _contexto.Usuarios.Where(x => x.TipoUsuario.idTipoUsuario == idTipoUsuario);
+
+            //Remove pedidos do usuário
+            IEnumerable<Pedido> pedidosRemover = _contexto.Pedidos.Where(x => x.Usuario.TipoUsuario.idTipoUsuario == idTipoUsuario).ToList();
+            _contexto.Pedidos.RemoveRange(pedidosRemover);
+            
+            //Remove endereço do usuário
+            IEnumerable<Endereco> enderecosRemover = _contexto.Enderecos.Where(x => x.Usuario.TipoUsuario.idTipoUsuario == idTipoUsuario).ToList();
+            _contexto.Enderecos.RemoveRange(enderecosRemover);
+
+            //Remove fichas de produção do usuário
+            IEnumerable<FichaProducao> fichasProducaoRemover = _contexto.FichasProducao.Where(x => x.Usuario.TipoUsuario.idTipoUsuario == idTipoUsuario).ToList();
+            if (fichasProducaoRemover.Count() > 0)
+                _contexto.FichasProducao.RemoveRange(fichasProducaoRemover);
+
             _contexto.RemoveRange(usuariosRemover);
             _contexto.SaveChanges();
-            return true;
         }
 
-        public bool RemoverPersonalizado(int id)
+        public void RemoverPersonalizado(int id)
         {
-            Usuario usuariosRemover = _contexto.Usuarios.First(x => x.idUsuario == id);
+            Usuario usuarioRemover = _contexto.Usuarios.First(x => x.idUsuario == id);
+            
             //Remove pedidos do usuário
             IEnumerable<Pedido> pedidosRemover = _contexto.Pedidos.Where(x => x.Usuario.idUsuario == id).ToList();
             _contexto.Pedidos.RemoveRange(pedidosRemover);
+            
             //Remove endereço do usuário
-            Endereco enderecoRemover = _contexto.Enderecos.First(x => x.Usuario.idUsuario == id);
-            _contexto.Enderecos.Remove(enderecoRemover);
+            IEnumerable<Endereco> enderecosRemover = _contexto.Enderecos.Where(x => x.Usuario.idUsuario == id).ToList();
+            _contexto.Enderecos.RemoveRange(enderecosRemover);
+            
             //Remove fichas de produção do usuário
             IEnumerable<FichaProducao> fichasProducaoRemover = _contexto.FichasProducao.Where(x => x.Usuario.idUsuario == id).ToList();
-            _contexto.FichasProducao.RemoveRange(fichasProducaoRemover);
+            if(fichasProducaoRemover.Count() > 0)
+                _contexto.FichasProducao.RemoveRange(fichasProducaoRemover);
+            
             //Remove o usuário
-            _contexto.Remove(usuariosRemover);
+            _contexto.Remove(usuarioRemover);
             _contexto.SaveChanges();
-            return true;
         }
     }
 }
